result/requirements.txt
opencv-python>=4.8.0
numpy>=1.21.0
python-dotenv>=0.21.0
redis>=4.5.0
aiohttp>=3.8.0
